Starting to convert the votes component to a directive
A directive that simply loads a hard-coded template file to represent the votes component.

The hard-coding needs replacing with scope variables, isolated to the directive.

The local express js app will serve assets in the same way the Rails asset pipeline will.  This is so that templates can be loaded from a URL and there will be no gotchas here.  In future, this can be optimised so that the templates are pre-compiled by the time they hit the pipeline, thus saving AJAX calls.

The tests need to be able to load the template files as well, in order to work.  I haven't figured out how to do this yet.

cc @iHiD