Ignoring max line length until the HTML strings move to files
The CoffeeScript multiline strings preserve whitespace which is causing issues (e.g. spaces in long URLs) when keeping to the max line length limit.

This limit will be enforced again when there are no more lengthy HTML strings in the code, i.e. they are all extracted to files of their own.