Conditionally load the correct template based on vote type
If the vote is for the item on the page the load the pageVote.html template.  If it is for a comment on the page load the commentVote.html

The templates are also cached as part of a grunt task which creates a separate module - meducationTemplates that is included as part of the meducationFrontEnd module.  This means no more AJAX calls in the tests or the production code.