Get upload working in browsers < IE 10
IE 8 and 9 need to use iframe transport for file uploads and so the data type is set to null as there is no meaningful type to set when going down this route.

The external_file_url field to send back to the Meducation server is now calculated from the response url + key in the form + the filename uploaded.